fix(lib): update MCP client, tool proxy, quality metrics, consent enforcement
Add result caching for read-only MCP tool calls (30s TTL), extend tool schema cache TTL to 5 min, add mutating-tool detection to bypass cache, and update tool proxy and consent enforcement logic.

+++ src/lib/quality-metrics.ts
@@ -52,7 +52,7 @@ export async function aggregateDailyMetrics(): Promise<void> {
     const sentCounts = await db
       .select({
         profileId: rainbowMessages.profileId,
-        count: sql<number>`count(*)::int`,
+        count: sql<number>`count(*)`,
       })
       .from(rainbowMessages)
       .where(
